fix: regenerate stale workflow to take advantage of new options
Newer versions of Projen automatically configure this to add
support for a "backlog" label on which this workflow won't run.
This allows us to avoid situations where a maintainer has to
continually remove the "stale" label on issues we consider valid
but haven't had time to work on yet.

This also adds a lock workflow in keeping with best practices from
our other open source projects.