chore: do not update examples as part of script
This does not currently work because the examples have `@cdktf/aws-cdk` as a dependency which have a peer dependency on specific versions of `cdktf` and thus subsequent installs of `yarn install` / `npm install` complain. However, now that we have Dependabot set up to manage dependency updates for the examples, Dependabot will create PRs to update these.