chore: auto-close PRs with no changes (#235)
#211 finally fixed the `provider-upgrade` workflow that hadn't been
working in months, but now a new problem appears: as you can see from
#233, we're now getting PRs that wind up empty after the self-mutation
step. In the short term, the best option is to add some steps to the build
workflow to check whether all of the changes were undone after Projen's
self-mutation, and to close the PR if so.