chore(ci): attempt to fix token permission issues (#914)
Something about [Projen
v0.87](https://github.com/projen/projen/releases/tag/v0.87.0) caused our
workflows to stop working with Dependabot. Specifically, the Checkout
step on the self-mutation job is failing with `Error: Input required and
not supplied: token`.

I think this is because Dependabot does not have access to secrets like
`PROJEN_GITHUB_TOKEN`, so the solution is to use the built-in
`GITHUB_TOKEN` and making sure that one has the appropriate permissions
configured. I'm not 100% positive this will work, but there's no way to
find out without merging this and trying it.