chore(deps): address security vulnerabilities (#1001)
There's a bunch of security alerts in this repo that can't/won't be
fixed until CDKTF comes out with a 0.20.12 or even a 0.21.0 release.
These all involve packages that are several layers deep (dependencies of
dependencies of dependencies) to the point where it's quite possible
this functionality really isn't used at all, but unfortunately the
security scanners aren't smart enough to pick that up. The easiest
(hacky but working) fix is to take advantage of Node's
`packageResolutions` to hard-code these to the fixed versions. These
overrides should be removed when we upgrade CDKTF.